Add detailed example for adding filters to README
While trying to remove extra fields from the Slack notification, we had to inspect the codebase to figure out the structure of `$notification_pieces`, and had trouble figuring out just where to place the filter code. I think this explanation in the README would point new users in the right direction.