Alert unpublished posts that go to a pending status as "unpublished"
Having a new post appear as "pending" is a very different situation than a post that is already published and that gets unpublished because the author makes a modification (so it's pending review to be republished). We use the "Event Manager" plugin that has this mechanism and the alert for an updated event (unpublished until approved" is much more urgent than a new event being submitted for review.